Make Microsoft.CodeAnalysis dependencies private assets
This ensures code analysis libraries - which are only needed
during development - don't get published and exposed to other
projects that consume this package.

Without this, other projects that consume this package and use
use a different version of the code analysis libraries will
run into dependency clashes.

This is because the the code analysis libraries have strict dependencies.
For instance, "Microsoft.CodeAnalysis/3.7.0" has a _strict_ dependency on
"Microsoft.CodeAnalysis.CSharp.Workspaces": "[3.7.0]", and is
thus incompatible with later versions.

https://docs.microsoft.com/en-us/nuget/consume-packages/package-references-in-project-files#controlling-dependency-assets